pyproject.toml
[tool.poetry]
name = "fastapiproject5"
version = "0.1.0"
description = ""
authors = ["mamdoot222 <a.h.sarohy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
openai = "^1.44.1"
sqlglot = "^25.20.1"
groq = "^0.11.0"
passlib = "^1.7.4"
motor = "^3.5.1"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
